Function of x on a fourier basis

Function of x on a fourier basis with a subset of covariates having a causal effect on Y using the parameters beta. The function is given by: $$f(x_i) = \sum_{j = 1}^p 1_{j \in js} \sum_{k = 1}^K (\beta_{j, k}^{(1)} \cos(0.2 k x_j) + \beta_{j, k}^{(2)} \sin(0.2 k x_j))$$

f_four(x, beta, js)

Arguments

x

a vector of covariates

beta

the parameter vector for the function f(X)

js

the indices of the causal covariates in X

Value

the value of the function f(x)

See also

Author

Markus Ulmer

Examples

set.seed(42)
# simulation of confounded data
sim_data <- simulate_data_nonlinear(q = 2, p = 150, n = 100, m = 2)
X <- sim_data$X
j <- sim_data$j[1]
apply(X, 1, function(x) f_four(x, sim_data$beta, j))
#>   [1] 0.9415741 0.9949555 1.1928927 1.1045218 0.9978997 0.9842344 0.4976803
#>   [8] 2.1840933 0.4847541 0.5049246 0.7075039 0.4811739 0.4798616 0.0132498
#>  [15] 0.4685555 0.8618920 0.9100779 0.9383755 0.5105662 0.4674675 0.5124398
#>  [22] 0.7417179 0.6072482 0.4155717 0.9032344 0.5803062 2.0788020 0.7597860
#>  [29] 2.0317824 1.4987971 2.0987924 0.4697440 0.7438867 2.1089224 0.6846367
#>  [36] 0.4675041 0.9439403 0.5775967 1.7517872 1.2243694 0.4981572 1.0936141
#>  [43] 1.4442143 0.5555504 1.4221857 0.8629139 0.8926297 0.5154851 2.1685337
#>  [50] 2.0593199 0.4679027 0.5537919 1.8799378 0.4691718 0.7381815 1.4696130
#>  [57] 0.5392643 0.9605256 1.9820476 0.5385648 1.0371837 1.9850954 0.4949520
#>  [64] 2.1623586 1.4495415 1.7156699 0.9397778 1.4580680 0.9223384 2.0135985
#>  [71] 1.6319588 0.7412391 0.8642937 1.9935582 2.1858894 1.4998491 1.6467786
#>  [78] 0.6237905 0.7634711 1.9498794 0.9000813 0.7788061 1.4230233 2.1884498
#>  [85] 0.5065487 0.9003522 1.5013830 0.5087280 0.9803233 0.5181828 0.6915481
#>  [92] 1.3673145 1.4464642 1.1582883 2.1166655 0.4860677 0.8651567 1.3980742
#>  [99] 1.8672286 0.7530638