Predictions of individual component functions for SDAM

Predicts the contribution of an individual component j using a fitted SDAM.

predict_individual_fj(object, j, x = NULL)

Arguments

object

Fitted object of class SDAM.

j

Which component to evaluate.

x

New numeric data to predict for.

Value

A vector of predictions for fj evaluated at Xjnew.

See also

Author

Cyrill Scheidegger

Examples

set.seed(1)
X <- matrix(rnorm(10 * 5), ncol = 5)
Y <- sin(X[, 1]) -  X[, 2] + rnorm(10)
model <- SDAM(x = X, y = Y, Q_type = "trim", trim_quantile = 0.5, nfold = 2, n_K = 1)
#> [1] "Initial cross-validation"
#> [1] "Second stage cross-validation"
predict_individual_fj(model, j = 1, seq(-2, 2, length.out = 100))
#>   [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [75] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0